auth_server_flask/requirements.txt
Flask>=2.0,<3.1
google-auth-oauthlib>=0.5,<1.3
google-cloud-secret-manager>=2.0,<2.19 # ★これを含める★
PyJWT>=2.0,<2.9
python-dotenv>=0.15,<1.1
functions_framework # functions-framework はローカルFlask実行では不要